+++ tap_mambu/transform.py
@@ -42,29 +42,29 @@ def remove_custom_nodes(this_json):
         if not kk[:1] == '_'}
 
 
+def add_cust_field(key, record, cust_field_sets):
+    for cf_key, cf_value in record.items():
+        field = {
+            'field_set_id' : key,
+            'id' : cf_key,
+            'value' : cf_value,
+        }
+        cust_field_sets.append(field)
+
 # Convert custom fields and sets
 # Generalize/Abstract custom fields to key/value pairs
 def convert_custom_fields(this_json):
-    new_json = this_json
-    i = 0
     for record in this_json:
         cust_field_sets = []
-        for key in record:
-            if isinstance(record[key], dict):
-                if key[:1] == '_':
-                    cust_field_set = {}
-                    cust_field_set['customFieldSetId'] = key
-                    cust_field_set_fields = []
-                    for cf_key, cf_value in record[key].items():
-                        field = {}
-                        field['customFieldId'] = cf_key
-                        field['customFieldValue'] = cf_value
-                        cust_field_set_fields.append(field)
-                    cust_field_set['customFieldValues'] = cust_field_set_fields
-                    cust_field_sets.append(cust_field_set)
-        new_json[i]['customFieldSets'] = cust_field_sets
-        i = i + 1
-    return new_json
+        for key, value in record.items():
+            if key.startswith('_'):
+                if isinstance(value, dict):
+                    add_cust_field(key, value, cust_field_sets)
+                elif isinstance(value, list):
+                    for element in value:
+                        add_cust_field(key, element, cust_field_sets)
+        record['custom_fields'] = cust_field_sets
+    return this_json
 
 
 # Run all transforms: denests _embedded, removes _embedded/_links, and